requirements.txt
streamlit
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.8
openai
faiss-cpu
tiktoken
python-dotenv
watchdog
pyyaml
numpy
pytest